Add Radio & Refactor Decryptor
This update finally adds the other radios from deezer. As of now there are 3 versions.
The "Flow Radio" used on the Home Tab.
The "Track Radio" is used when you start a radio from a track or simply play a single track.
And the "Playlist/Album Radio" this one makes sure that the playlist/album doesn't end after the last track was played.

As for the decryptor, there was a problem which made flac quality tracks skip or go silent while listening. This should be fixed or at least shouldn't occur as often